I think I'm finished with Section D, haven't added optional feature.

+++ ITEC145-Section-D/Form1.cs
@@ -9,9 +9,6 @@ namespace ITEC145_Section_D
         List<cars> allCars = new List<cars>();
 
         string projectDirectory = Directory.GetCurrentDirectory();      //Got this idea from JJ :D
-
-
-
     struct cars
         {
             public string Make;
@@ -19,8 +16,6 @@ namespace ITEC145_Section_D
             public int Year;
             public double Mileage;
         }
-        
-
         public Form1()
         {
             InitializeComponent();
@@ -30,11 +25,11 @@ namespace ITEC145_Section_D
                 dataGridCars.Columns[i].HeaderText = columnNames[i];    //Sets each columns name based on the names in columnNames
             }
         }
-
         private void btnLoad_Click(object sender, EventArgs e)
         {
             OpenFileDialog ofd = new OpenFileDialog();
             ofd.InitialDirectory = projectDirectory;
+
             try
             {
                 if(ofd.ShowDialog() == DialogResult.OK)                     //If file dialog is successful in selecting a file. Do something.
@@ -48,11 +43,10 @@ namespace ITEC145_Section_D
                         cars carList = new cars();
                         words = reader.ReadLine().Split(",");
                     
-                        carList.Make = words[0];                    
-                        carList.Model = words[1];                    
-                        carList.Year = int.Parse(words[2]);        
-                        carList.Mileage = double.Parse(words[3]);
-
+                        carList.Make = words[0];                            //Would like to figure out how to loop through a struct,
+                        carList.Model = words[1];                           //to figure out how many variables there are in order
+                        carList.Year = int.Parse(words[2]);                 //to make this code more responsive based on the column
+                        carList.Mileage = double.Parse(words[3]);           //names you give it above.
                         allCars.Add(carList);
                         count++;
                     }
@@ -72,13 +66,15 @@ namespace ITEC145_Section_D
 
         private void btnLoadGrid_Click(object sender, EventArgs e)
         {
-            for(int i = 0; i < allCars.Count; i++)
+            int actualRowCount = dataGridCars.RowCount - 1;     //Used to add multiple records to datagrid.
+
+            for(int i = 0 + actualRowCount; i < allCars.Count + actualRowCount; i++)
             {
                 dataGridCars.Rows.Add();
-                dataGridCars.Rows[i].Cells[0].Value = allCars[i].Make;
-                dataGridCars.Rows[i].Cells[1].Value = allCars[i].Model;
-                dataGridCars.Rows[i].Cells[2].Value = allCars[i].Year;
-                dataGridCars.Rows[i].Cells[3].Value = allCars[i].Mileage;
+                dataGridCars.Rows[i].Cells[0].Value = allCars[i - actualRowCount].Make;         //Again would like to make this more responsive
+                dataGridCars.Rows[i].Cells[1].Value = allCars[i - actualRowCount].Model;        //to a single global field change
+                dataGridCars.Rows[i].Cells[2].Value = allCars[i - actualRowCount].Year;         //ran out of time to figure this out.
+                dataGridCars.Rows[i].Cells[3].Value = allCars[i - actualRowCount].Mileage;
             }
         }
 
@@ -88,43 +84,50 @@ namespace ITEC145_Section_D
             {
                 int count = 0;
                 string filename = txtFilename.Text;
-                allCars.Clear();
-                cars newCars = new cars();
 
-                for (int i = 0; i < dataGridCars.RowCount - 1; i++)
+                if(filename == "")
                 {
-                    newCars.Make = dataGridCars.Rows[i].Cells[0].Value.ToString();
-                    newCars.Model = dataGridCars.Rows[i].Cells[1].Value.ToString();
-                    newCars.Year = int.Parse(dataGridCars.Rows[i].Cells[2].Value.ToString());
-
-                    if (newCars.Year > DateTime.Now.Year+1)
-                    {
-                        MessageBox.Show("You have a car in the datagrid that doesnt exist yet. Filesave operation cancelled.");
-                        throw new Exception();
-                    }
-                    else if(newCars.Year < DateTime.Now.Year - 100 && count < 1)
-                    {
-                        count++;
-                        MessageBox.Show("A car in your datagrid is likely too old to exist, advised to check your data and save again. This will not show again");
-                    }
-
-                    newCars.Mileage = double.Parse(dataGridCars.Rows[i].Cells[3].Value.ToString());
-                    allCars.Add(newCars);
+                    MessageBox.Show("Please name your file!");
                 }
-
-                StreamWriter writer = new StreamWriter(filename + ".txt");      //Placed here to prevent creation of a file in the event of an error
-
-                foreach (cars cars in allCars)
+                else
                 {
-                    writer.WriteLine(cars.Make + "," + cars.Model + "," + cars.Year + "," + cars.Mileage);
-                }
-                writer.Close();
+                    allCars.Clear();
+                    cars newCars = new cars();
 
-                MessageBox.Show($"{filename}.txt saved succesfully!");
+                    for (int i = 0; i < dataGridCars.RowCount - 1; i++)
+                    {
+                        newCars.Make = dataGridCars.Rows[i].Cells[0].Value.ToString();
+                        newCars.Model = dataGridCars.Rows[i].Cells[1].Value.ToString();
+                        newCars.Year = int.Parse(dataGridCars.Rows[i].Cells[2].Value.ToString());
+
+                        if (newCars.Year > DateTime.Now.Year+1)
+                        {
+                            MessageBox.Show("You have a car in the datagrid that doesnt exist yet. Filesave operation cancelled.");
+                            throw new Exception();
+                        }
+                        else if(newCars.Year < DateTime.Now.Year - 100 && count < 1)
+                        {
+                            count++;
+                            MessageBox.Show("A car in your datagrid is likely too old to exist, advised to check your data and save again. This will not show again");
+                        }
+
+                        newCars.Mileage = double.Parse(dataGridCars.Rows[i].Cells[3].Value.ToString());
+                        allCars.Add(newCars);
+                    }
+
+                    StreamWriter writer = new StreamWriter(filename + ".txt");      //Placed here to prevent creation of a file in the event of an error
+
+                    foreach (cars cars in allCars)
+                    {
+                        writer.WriteLine(cars.Make + "," + cars.Model + "," + cars.Year + "," + cars.Mileage);
+                    }
+                    writer.Close();
+                    MessageBox.Show($"{filename}.txt saved succesfully!");
+                }
             }
             catch
             {
-                MessageBox.Show("Please check that you have saved only integers and doubles in their respective fields (Year and Mileage), and try again. Filesave was cancelled.");
+                MessageBox.Show("Something went wrong. Please check that you have saved only integers and doubles in their respective fields (Year and Mileage), and try again. Filesave was cancelled.");
             }
             
         }
